Remove the rest of the implict conversions; everything now explicit where needed.

+++ src/Cortex.BeaconNode.Api/BeaconNodeApiAdapter.cs
@@ -61,7 +61,7 @@ namespace Cortex.BeaconNode.Api
 
             var result = new BeaconBlock()
             {
-                Slot = data.Slot,
+                Slot = (ulong)data.Slot,
                 Body = new BeaconBlockBody()
                 {
                     Randao_reveal = data.Body!.RandaoReveal.AsSpan().ToArray()
